api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
sqlalchemy>=2.0.0
alembic>=1.14.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
stripe>=9.0.0
google-auth-oauthlib>=1.2.0
httpx>=0.27.0
slowapi>=0.1.9
